@@ -15,7 +15,7 @@
  */
 package com.vaadin.flow.spring.test;
 
-import org.junit.Assert;
+import org.junit.jupiter.api.Assertions;
 import org.junit.Test;
 import org.openqa.selenium.By;
 import org.openqa.selenium.WebElement;
@@ -32,18 +32,18 @@ public class CoExistingSpringEndpointsIT extends AbstractSpringTest {
         open();
 
         String nonExistingRoutePath = "non-existing-route";
-        Assert.assertTrue(isElementPresent(By.id("main")));
+        Assertions.assertTrue(isElementPresent(By.id("main")));
 
         getDriver().get(getContextRootURL() + '/' + nonExistingRoutePath);
 
-        Assert.assertTrue(getDriver().getPageSource().contains(String
+        Assertions.assertTrue(getDriver().getPageSource().contains(String
                 .format("Could not navigate to '%s'", nonExistingRoutePath)));
-        Assert.assertTrue(getDriver().getPageSource().contains(String.format(
+        Assertions.assertTrue(getDriver().getPageSource().contains(String.format(
                 "Reason: Couldn't find route for '%s'", nonExistingRoutePath)));
 
         getDriver().get(getContextRootURL() + "/oauth/authorize");
 
         WebElement header = findElement(By.tagName("h1"));
-        Assert.assertEquals("OAuth Error", header.getText());
+        Assertions.assertEquals("OAuth Error", header.getText());
     }
 }
