@@ -15,7 +15,7 @@
  */
 package com.vaadin.flow.spring.test;
 
-import org.junit.Assert;
+import org.junit.jupiter.api.Assertions;
 import org.junit.Test;
 
 import com.vaadin.testbench.TestBenchElement;
@@ -28,7 +28,7 @@ public class ComponentAddedViaInitListenerIT extends AbstractSpringTest {
 
         TestBenchElement component = $("init-listener-component").first();
         TestBenchElement div = component.$("div").first();
-        Assert.assertEquals("Init Listener Component", div.getText());
+        Assertions.assertEquals("Init Listener Component", div.getText());
     }
 
     @Override
