@@ -17,7 +17,7 @@ package com.vaadin.flow.spring.test;
 
 import java.util.List;
 
-import org.junit.Assert;
+import org.junit.jupiter.api.Assertions;
 import org.junit.Test;
 
 import com.vaadin.testbench.TestBenchElement;
@@ -31,16 +31,16 @@ public class DoubleNpmAnnotationIT extends AbstractSpringTest {
         List<TestBenchElement> paperInputs = $("paper-input").all();
 
         // check that elements are on the page
-        Assert.assertTrue("Should have found a 'paper-checkbox'",
+        Assertions.assertTrue("Should have found a 'paper-checkbox'",
                 paperCheckboxes.size() > 0);
-        Assert.assertTrue("Should have found a 'paper-input'",
+        Assertions.assertTrue("Should have found a 'paper-input'",
                 paperInputs.size() > 0);
 
         // verify that the paper components are upgraded
-        Assert.assertNotNull(
+        Assertions.assertNotNull(
                 "'paper-checkbox' should have had element in shadow dom",
                 paperCheckboxes.get(0).$("checkboxContainer"));
-        Assert.assertNotNull(
+        Assertions.assertNotNull(
                 "'paper-input' should have had element in shadow dom",
                 paperInputs.get(0).$("paper-input-container"));
     }
