@@ -1,6 +1,6 @@
 package com.vaadin.flow.spring.test;
 
-import org.junit.Assert;
+import org.junit.jupiter.api.Assertions;
 import org.junit.Test;
 import org.junit.experimental.categories.Category;
 
@@ -15,25 +15,25 @@ public class ResourcesIT extends AbstractSpringTest {
 
     @Test
     public void resourceInPublic() {
-        Assert.assertEquals("This is in the public folder on the classpath",
+        Assertions.assertEquals("This is in the public folder on the classpath",
                 loadFile("/public-file.txt"));
     }
 
     @Test
     public void resourceInStatic() {
-        Assert.assertEquals("This is in the static folder on the classpath",
+        Assertions.assertEquals("This is in the static folder on the classpath",
                 loadFile("/static-file.txt"));
     }
 
     @Test
     public void resourceInResources() {
-        Assert.assertEquals("This is in the resources folder on the classpath",
+        Assertions.assertEquals("This is in the resources folder on the classpath",
                 loadFile("/resources-file.txt"));
     }
 
     @Test
     public void resourceInMetaInfResources() {
-        Assert.assertEquals(
+        Assertions.assertEquals(
                 "This is in the META-INF/resources folder on the classpath",
                 loadFile("/metainfresources-file.txt"));
     }
