@@ -21,7 +21,7 @@ import java.io.FileWriter;
 import java.io.IOException;
 import java.util.List;
 
-import org.junit.Assert;
+import org.junit.jupiter.api.Assertions;
 import org.junit.Ignore;
 import org.junit.Test;
 import org.openqa.selenium.By;
@@ -51,7 +51,7 @@ public class UploadIT extends AbstractSpringTest {
 
         waitUntil(driver -> isElementPresent(By.className("uploaded-text")));
         WebElement uploadedText = findElement(By.className("uploaded-text"));
-        Assert.assertEquals("foo", uploadedText.getText());
+        Assertions.assertEquals("foo", uploadedText.getText());
     }
 
     @Test
@@ -63,7 +63,7 @@ public class UploadIT extends AbstractSpringTest {
 
         List<TestBenchElement> inputs = $(TestBenchElement.class).id("upload")
                 .$("input").all();
-        Assert.assertFalse(
+        Assertions.assertFalse(
                 "Upload element is not initialized: it doesn't contain "
                         + "any child element (so it has no element in the shadow root)",
                 inputs.isEmpty());
