@@ -15,8 +15,8 @@
  */
 package com.vaadin.flow.spring;
 
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletRequestWrapper;
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletRequestWrapper;
 
 import org.springframework.web.util.UrlPathHelper;
 
