@@ -18,8 +18,8 @@ package com.vaadin.flow.spring;
 import java.util.Arrays;
 import java.util.Collections;
 
-import org.junit.Assert;
-import org.junit.Test;
+import org.junit.jupiter.api.Assertions;
+import org.junit.jupiter.api.Test;
 
 public class PrefixTreeTest {
 
@@ -27,25 +27,25 @@ public class PrefixTreeTest {
     public void hasPrefix_containsPrefix_returnsTrue() {
         PrefixTree prefixTree = new PrefixTree(
                 Arrays.asList("com/sun", "antlr", "ch/quos/logback"));
-        Assert.assertTrue(prefixTree.hasPrefix("antlr"));
-        Assert.assertTrue(prefixTree.hasPrefix("com/sun/test"));
-        Assert.assertTrue(prefixTree.hasPrefix("com/sun"));
+        Assertions.assertTrue(prefixTree.hasPrefix("antlr"));
+        Assertions.assertTrue(prefixTree.hasPrefix("com/sun/test"));
+        Assertions.assertTrue(prefixTree.hasPrefix("com/sun"));
     }
 
     @Test
     public void hasPrefix_doesNotContainPrefix_returnsFalse() {
         PrefixTree prefixTree = new PrefixTree(
                 Arrays.asList("com/sun", "antlr", "ch/quos/logback"));
-        Assert.assertFalse(prefixTree.hasPrefix(""));
-        Assert.assertFalse(prefixTree.hasPrefix("a"));
-        Assert.assertFalse(prefixTree.hasPrefix("test"));
-        Assert.assertFalse(prefixTree.hasPrefix("com/su"));
+        Assertions.assertFalse(prefixTree.hasPrefix(""));
+        Assertions.assertFalse(prefixTree.hasPrefix("a"));
+        Assertions.assertFalse(prefixTree.hasPrefix("test"));
+        Assertions.assertFalse(prefixTree.hasPrefix("com/su"));
     }
 
     @Test
     public void hasPrefix_emptyTree_returnsFalse() {
         PrefixTree prefixTree = new PrefixTree(Collections.emptyList());
-        Assert.assertFalse(prefixTree.hasPrefix("a"));
+        Assertions.assertFalse(prefixTree.hasPrefix("a"));
     }
 
 }
