@@ -18,9 +18,9 @@ package com.vaadin.flow.spring.scopes;
 import java.util.Properties;
 import java.util.concurrent.atomic.AtomicInteger;
 
-import org.junit.After;
-import org.junit.Assert;
-import org.junit.Test;
+import org.junit.jupiter.api.AfterEach;
+import org.junit.jupiter.api.Assertions;
+import org.junit.jupiter.api.Test;
 import org.mockito.Mockito;
 import org.springframework.beans.factory.ObjectFactory;
 import org.springframework.beans.factory.config.Scope;
@@ -49,26 +49,26 @@ public abstract class AbstractScopeTest {
 
     }
 
-    @After
+    @AfterEach
     public void clearSession() {
         session = null;
     }
 
-    @Test(expected = IllegalStateException.class)
+    @Test
     public void get_noCurrentSession_throwException() {
         Scope scope = getScope();
 
         scope.get("foo", Mockito.mock(ObjectFactory.class));
     }
 
-    @Test(expected = IllegalStateException.class)
+    @Test
     public void registerDestructionCallback_noCurrentSession_throwException() {
         Scope scope = getScope();
 
         scope.registerDestructionCallback("foo", Mockito.mock(Runnable.class));
     }
 
-    @Test(expected = IllegalStateException.class)
+    @Test
     public void remove_noCurrentSession_throwException() {
         Scope scope = getScope();
 
@@ -84,14 +84,14 @@ public abstract class AbstractScopeTest {
         when(factory.getObject()).thenReturn(object);
         Object fooObject = scope.get("foo", factory);
 
-        Assert.assertSame(object, fooObject);
+        Assertions.assertSame(object, fooObject);
         verify(factory).getObject();
 
         fooObject = scope.get("foo", factory);
 
         // This time it has not been called. Otherwise this would have failed.
         verify(factory).getObject();
-        Assert.assertSame(object, fooObject);
+        Assertions.assertSame(object, fooObject);
     }
 
     @SuppressWarnings("rawtypes")
@@ -121,7 +121,7 @@ public abstract class AbstractScopeTest {
 
         scope.getBeanStore().destroy();
 
-        Assert.assertEquals(1, count.get());
+        Assertions.assertEquals(1, count.get());
     }
 
     @SuppressWarnings("unchecked")
